List klippy dependencies

This function is used to get the list of klippy dependencies.

klippy_dependencies()

Value

A list of dependencies that can be passed to attachDependencies or html_document_base.

See also

Other HTML dependencies functions: html-dependencies